feat: ask user wheather auto install node_modules
